chore: rename Runnable to WorldTimer
